Separated Project into modules
