--- src/Bejeweled/module-info.java
@@ -1,7 +1,0 @@
-module Bejeweled{
-	exports Bejeweled;
-	requires TMGE;
-	uses TMGE;
-	uses Player;
-	provides BejeweledTile with Tile;
-}
