+++ src/Bejeweled/src/module-info.java
@@ -1,0 +1,5 @@
+module Bejeweled {
+	exports Bejeweled;
+	requires java.desktop;
+	requires TMGE;
+}
