+++ src/GUI/src/module-info.java
@@ -1,0 +1,5 @@
+module GUI {
+	requires java.desktop;
+	requires Bejeweled;
+	requires Memory;
+}
