--- src/Memory/module-info.java
@@ -1,10 +1,0 @@
-import TMGE.*;
-
-module Memory{
-	exports Memory;
-	requires TMGE;
-	uses TMGE;
-	uses Player;
-	provides MemoryTile with Tile;
-
-}
