+++ src/Memory/src/module-info.java
@@ -1,0 +1,5 @@
+module Memory {
+	requires java.desktop;
+	requires TMGE;
+	exports Memory;
+}
