--- src/TMGE/module-info.java
@@ -1,9 +1,0 @@
-module TMGE{
-	exports Player;
-	exports Tile;
-	exports TMGE;
-	requires Bejeweled;
-	requires MemoryMatching;
-	uses Bejeweled;
-	uses Memory;
-}
