+++ src/TMGE/src/module-info.java
@@ -1,0 +1,4 @@
+module TMGE {
+	exports TMGE;
+	requires java.desktop;
+}
